pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "autolab"
version = "1.1.19"
description = "Shared autolab CLI package for workflow orchestration"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["autolab", "workflow", "experiment", "orchestration"]
dependencies = [
  "PyYAML>=6",
  "jsonschema>=4.18.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff>=0.15.0",
  "mdformat>=1.0.0",
  "yamlfix>=1.19.0",
]

[project.scripts]
autolab = "autolab.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
autolab = [
  "scaffold/.autolab/experiment_file_line_limits.yaml",
  "scaffold/.autolab/verifier_policy.yaml",
  "scaffold/.autolab/prompts/*.md",
  "scaffold/.autolab/prompts/shared/*.md",
  "scaffold/.autolab/prompts/shared/.gitkeep",
  "scaffold/.autolab/policy/*.yaml",
  "scaffold/.autolab/schemas/*.json",
  "scaffold/.autolab/verifiers/*.py",
  "skills/*/*/*.md",
]
